Set amount of tickets to int instead of float

+++ src/main/java/ch/wisv/areafiftylan/products/model/order/Order.java
@@ -94,7 +94,7 @@ public class Order {
     }
 
     @JsonView(View.OrderOverview.class)
-    public float getAmountTickets() {
+    public int getAmountTickets() {
         return tickets.size();
     }
